Changed configuration method, moved to Spring DI. Contains moshi problem
