+++ pom.xml
@@ -41,13 +41,13 @@
 		<dependency>
 			<groupId>com.squareup.moshi</groupId>
 			<artifactId>moshi-kotlin</artifactId>
-			<version>1.5.0</version>
+			<version>1.10.0</version>
 		</dependency>
 		<!-- https://mvnrepository.com/artifact/com.squareup.moshi/moshi-adapters -->
 		<dependency>
 			<groupId>com.squareup.moshi</groupId>
 			<artifactId>moshi-adapters</artifactId>
-			<version>1.5.0</version>
+			<version>1.10.0</version>
 		</dependency>
 		<!-- https://mvnrepository.com/artifact/com.squareup.okhttp3/okhttp -->
 		<dependency>
@@ -87,6 +87,17 @@
 			<groupId>org.mnode.ical4j</groupId>
 			<artifactId>ical4j</artifactId>
 			<version>2.1.5</version>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-configuration-processor</artifactId>
+			<optional>true</optional>
+		</dependency>
+		<!-- https://mvnrepository.com/artifact/com.squareup.okio/okio -->
+		<dependency>
+			<groupId>com.squareup.okio</groupId>
+			<artifactId>okio</artifactId>
+			<version>1.16.0</version>
 		</dependency>
 	</dependencies>
