--- src/main/kotlin/com/example/meetingBot/mailApi/IcsCalendar.kt
@@ -1,6 +1,0 @@
-package com.example.meetingBot.mailApi
-
-import org.springframework.stereotype.Component
-
-@Component
-class IcsCalendar(val content: ByteArray)
