--- src/main/kotlin/com/example/meetingBot/main/ZoomMeeting.kt
@@ -1,14 +1,0 @@
-package com.example.meetingBot.main
-
-import org.springframework.stereotype.Component
-import java.time.Duration
-import java.time.Instant
-
-class ZoomMeeting(
-    val topic: String?,
-    val agenda: String?,
-    val startTime: Instant,
-    val duration: Duration,
-    val joinUrl: String,
-    var attendees: List<String>?,
-)
