--- src/main/kotlin/com/example/meetingBot/zoomApi/infrastructure/RequestMethod.kt
@@ -1,8 +1,0 @@
-package com.example.meetingBot.zoomApi.infrastructure
-
-/**
- * Provides enumerated HTTP verbs
- */
-enum class RequestMethod {
-    GET, DELETE, HEAD, OPTIONS, PATCH, POST, PUT
-}
