+++ src/main/kotlin/qiwi/zoombot/spring/config/SchedulingConfiguration.kt
@@ -1,0 +1,12 @@
+package qiwi.zoombot.spring.config
+
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
+import org.springframework.context.annotation.Configuration
+import org.springframework.scheduling.annotation.EnableScheduling
+
+@Configuration
+@EnableScheduling
+//@ConditionalOnProperty(name = ["scheduling.enabled"], matchIfMissing = true)
+class SchedulingConfiguration{
+
+}
