+++ src/main/kotlin/qiwi/zoombot/spring/mailApi/IcsCalendar.kt
@@ -1,0 +1,3 @@
+package qiwi.zoombot.spring.mailApi
+
+class IcsCalendar(val content: ByteArray)
