+++ src/main/kotlin/qiwi/zoombot/spring/zoomApi/infrastructure/RequestMethod.kt
@@ -1,0 +1,8 @@
+package qiwi.zoombot.spring.zoomApi.infrastructure
+
+/**
+ * Provides enumerated HTTP verbs
+ */
+enum class RequestMethod {
+    GET, DELETE, HEAD, OPTIONS, PATCH, POST, PUT
+}
