--- src/test/kotlin/com/example/meetingBot/ZoomApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.example.meetingBot
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class ZoomApplicationTests {
-
-	@Test
-	fun contextLoads() {
-	}
-
-}
