+++ src/test/kotlin/qiwi/zoombot/spring/ZoomApplicationTests.kt
@@ -1,0 +1,13 @@
+package qiwi.zoombot.spring
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class ZoomApplicationTests {
+
+	@Test
+	fun contextLoads() {
+	}
+
+}
